Make cron_lib.php a class
