+++ classes/task/cron_langs.php
@@ -22,9 +22,13 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-
 namespace mod_bacs\task;
 
+defined('MOODLE_INTERNAL') || die();
+
+require_once(dirname(__FILE__, 2) . '/cron_lib.php');
+
+use mod_bacs\cron_lib;
 use core\session\exception;
 use core\task\scheduled_task;
 use dml_exception;
@@ -52,8 +56,7 @@ class cron_langs extends scheduled_task {
      * @throws dml_transaction_exception
      */
     public function execute() {
-        require_once(dirname(dirname(dirname(__FILE__))) . '/cron/cron_bacs/cron_lib.php');
 
-        cron_langs();
+        cron_lib::cron_langs();
     }
 }
