+++ classes/task/cron_tasks.php
@@ -22,9 +22,13 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-
 namespace mod_bacs\task;
 
+defined('MOODLE_INTERNAL') || die();
+
+require_once(dirname(__FILE__, 2) . '/cron_lib.php');
+
+use mod_bacs\cron_lib;
 use core\task\scheduled_task;
 use dml_exception;
 
@@ -48,8 +52,6 @@ class cron_tasks extends scheduled_task {
      * @throws dml_exception
      */
     public function execute() {
-        require_once(dirname(dirname(dirname(__FILE__))) . '/cron/cron_bacs/cron_lib.php');
-
-        cron_tasks();
+        cron_lib::cron_tasks();
     }
 }
